feat: render examples, token-budget output, path grouping; modularize markdown
Implements three enhancements and refactors the markdown generator.

Features:
- #6: --include-examples now renders request/response examples as fenced
  JSON blocks at --detail full, resolving $refs into components/examples
  (previously only a placeholder). Examples maps switched to IndexMap so
  rendered output stays deterministic.
- #7: --max-tokens fits output to a token budget, stepping the detail level
  down (full -> standard -> basic -> summary) until it fits and reporting any
  reduction on stderr.
- #8: --group-by path groups endpoints by path, one section per path with its
  methods listed underneath, in spec order.

Refactor (no behavior change):
- Split the ~1200-line markdown.rs into a markdown/ module (views, endpoint,
  schema, examples) behind the unchanged generate_markdown entry point.
- De-duplicate the view preamble and endpoint filters; centralize the HTTP
  method list as PathItem::operations and the JSON-pointer decode helper.

Adds integration tests and an examples fixture; all tests pass.
